M2-6493: rename RespondentData -> ParticipantActivity
- update routes to nest RespondentData under /:participantId
