+++ src/modules/Dashboard/features/ParticipantActivity/ParticipantActivity.styles.tsx
