--- src/modules/Dashboard/features/ParticipantActivityDetails/ParticipantActivityDetails.tsx
@@ -1,3 +1,0 @@
-export const ParticipantActivityDetails = () => <div>ParticipantActivityDetails feature</div>;
-
-export default ParticipantActivityDetails;
