--- src/modules/Dashboard/features/ParticipantActivityDetails/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ParticipantActivityDetails';
