--- src/modules/Dashboard/pages/ParticipantActivityDetails/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ParticipantActivityDetails';
