--- src/modules/Dashboard/pages/RespondentData/index.ts
@@ -1,3 +1,0 @@
-import { RespondentData } from './RespondentData';
-
-export default RespondentData;
