+++ src/modules/Dashboard/pages/index.ts
@@ -1,3 +1,2 @@
 export * from './Applet';
 export * from './Main';
-export * from './RespondentData';
